gomod: move shared test helpers to helpers.py
Move proc_mock, get_mock_dir, get_mocked_data into helpers.py
